it smoother

+++ JavaServer/src/main/java/util/NetworkUtils/GetUtils.java
@@ -10,8 +10,8 @@ public class GetUtils {
         double x = buffer.getDouble();
         double y = buffer.getDouble();
 
-        x = (double) Math.round(x*100)/100;
-        y = (double) Math.round(y*100)/100;
+        x = Math.round(x*100.0)/100.0;
+        y = Math.round(y*100.0)/100.0;
 
         return new Position(x, y);
     }
